added entity and entity id support

+++ infra/flow_source_sine.py
@@ -4,9 +4,9 @@ from domain.entities.flow_source import FlowSource
 
 
 class FlowSourceSine(FlowSource):
-    _current = 0.
+    _current = 0.0
     _step = 0.1
-    _freq: int
+    _freq: float
     _amplitude: float
 
     def __init__(self, freq: float, amplitude: float):
@@ -14,7 +14,8 @@ class FlowSourceSine(FlowSource):
         self._amplitude = amplitude
 
     def get_current(self) -> Flow:
-        flow = Flow(self._amplitude *
-                    math.sin(self._current * self._freq * 2 * math.pi))
+        flow = Flow(
+            self._amplitude * math.sin(self._current * self._freq * 2 * math.pi)
+        )
         self._current += self._step
         return flow
